requirements.txt
PyQt5==5.10.1
capstone==4.0.1
keystone==0.9.1
pycryptodome==3.8.0;platform_system=="Windows"
pycrypto==2.6.1;platform_system=="Linux"